requirements.txt
runpod>=1.0.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.0.0
safetensors>=0.3.0